requirements.txt
certifi==2019.11.28
pycryptodome==3.8.2
wincertstore==0.2
